Made VR360 accept zooming with the CTRL+ALT+RMB or with the mousewheel. Fixed permissions issues in DEB and added maintainer. Allowed support for sequences without a period, like image0001.exr. Improved window positioning when the image is too big to fit the desktop.
Former-commit-id: 68b345cd87f5533fa1967e015233620799f3b524
Former-commit-id: 2f13b3320f2b7e646cbd53c69157322e25012c9a

+++ shaders/YByRy.glsl
@@ -15,10 +15,10 @@ uniform sampler2D VImage;
 uniform sampler3D lut;
 
 // Interlaced/Checkerboard controls
-uniform float mask;
-uniform float mask_value;
-uniform float height;
-uniform float width;
+uniform int mask;
+uniform int mask_value;
+uniform int height;
+uniform int width;
 
 // Standard controls
 uniform float gain;
@@ -120,7 +120,7 @@ void main()
     }
 
 
-  float x = 1000.0;
+  int x = 1000;
   if ( mask == 1.0 )  // even odd rows
   {
       x = mod( gl_TexCoord[0].t * height, 2.0 );
